Rebased branch and updated the module name
